fix(iori): pipe using tokio

+++ crates/iori/src/merge/pipe.rs
@@ -9,7 +9,7 @@ use std::{
 
 use tokio::{
     fs::File,
-    sync::{Mutex, MutexGuard},
+    sync::{mpsc, Mutex, MutexGuard},
 };
 
 use crate::{error::IoriResult, StreamingSegment, ToSegmentData};
@@ -23,42 +23,55 @@ pub struct PipeMerger<S> {
     output_dir: PathBuf,
     recycle: bool,
 
+    sender: mpsc::UnboundedSender<S>,
     next: Arc<AtomicU64>,
     segments: Arc<Mutex<HashMap<u64, Option<S>>>>,
 }
 
-impl<S> PipeMerger<S> {
+impl<S> PipeMerger<S>
+where
+    S: StreamingSegment + Send + Sync + 'static,
+{
     pub fn new<P>(output_dir: P, recycle: bool) -> Self
     where
         P: Into<PathBuf>,
     {
+        let output_dir = output_dir.into();
+
+        let (sender, mut receiver) = mpsc::unbounded_channel::<S>();
+        let output_dir_receiver = output_dir.clone();
+        tokio::spawn(async move {
+            let stdout = &mut tokio::io::stdout();
+            while let Some(segment) = receiver.recv().await {
+                let path = segment_path(&segment, &output_dir_receiver);
+                let mut file = File::open(&path).await.unwrap();
+                tokio::io::copy(&mut file, stdout).await.unwrap();
+
+                if recycle {
+                    tokio::fs::remove_file(&path).await.unwrap();
+                }
+            }
+        });
+
         Self {
-            output_dir: output_dir.into(),
+            output_dir,
             recycle,
 
+            sender,
             next: Arc::new(AtomicU64::new(0)),
             segments: Arc::new(Mutex::new(HashMap::new())),
         }
     }
-}
 
-impl<S> PipeMerger<S>
-where
-    S: StreamingSegment + ToSegmentData + Send + Sync + 'static,
-{
     async fn pipe_segments(
         &self,
         mut segments: MutexGuard<'_, HashMap<u64, Option<S>>>,
     ) -> IoriResult<()> {
         while let Some(segment) = segments.remove(&self.next.load(Ordering::Relaxed)) {
             if let Some(segment) = segment {
-                let path = segment_path(&segment, &self.output_dir);
-                // open file and write binary content to stdout
-                let mut file = std::fs::File::open(&path)?;
-                let _ = std::io::copy(&mut file, &mut std::io::stdout());
-                if self.recycle {
-                    tokio::fs::remove_file(&path).await?;
-                }
+                let Ok(_) = self.sender.send(segment) else {
+                    break;
+                };
             }
 
             self.next.fetch_add(1, Ordering::Relaxed);
@@ -111,7 +124,10 @@ where
     }
 
     async fn finish(&mut self) -> IoriResult<Self::MergeResult> {
-        tokio::fs::remove_dir_all(&self.output_dir).await?;
+        if self.recycle {
+            tokio::fs::remove_dir_all(&self.output_dir).await?;
+        }
+
         Ok(())
     }
 }
